django/requirements.txt
asgiref==3.6.0
Django==4.2
django-ninja==0.21.0
gunicorn==20.1.0
psycopg2==2.9.6
pydantic==1.10.7
sqlparse==0.4.3
typing_extensions==4.5.0
